🎨 style(AndroidManifest.xml): add SplashScreen activity and reorder activities 🚚 chore(AndroidManifest.xml): remove unused drawables and reorder activities The SplashScreen activity has been added to the AndroidManifest.xml file, and the order of activities has been changed to match the order in which they appear in the app. Unused drawables have been removed from the project.
🎨 style(activity_encryption.xml): update ImageView dimensions and src
🎨 style(activity_main.xml): update ImageView src
🎨 style(activity_splash_screen.xml): add new layout file
🎨 style(ic_launcher.xml): update drawable resources
🎨 style(ic_launcher_round.xml): update drawable resources
🎨 style(mipmap-hdpi): add new ic_launcher.png file
🎨 style(mipmap-mdpi): add new ic_launcher.png file

The changes in the XML files update the app's UI by changing the dimensions and source of the ImageView elements. The new layout file adds a splash screen to the app. The changes in the mipmap resources add new image files to the app.

🎨 style(res): update app launcher icons and background color
The app launcher icons have been updated to use PNG format instead of WebP format. The ic_launcher_background.xml file has been added to define the background color of the app launcher icons. The windowIsTranslucent attribute has been added to the app theme to make the app launcher icons more visible on devices with a dark background.